add test for 3500 pbmcs
At this point it seems that harmonypy results are nearly identical to
the results produced by the harmony R package.

For 3500 cells (PBMCs), harmonypy finishes in 25 seconds versus 5
seconds for the harmony R package (written in Rcpp).

The average correlation coefficient between the R and Python results
0.99 for 30 adjusted PCs. This is great, and probably will not improve.

It may be worthwhile to consider approaches to speed up harmonypy:
- profiliing the code
- numba
- cython